backend/requirements.txt
# --- Core Web Framework ---
gunicorn>=21.2.0
uvicorn[standard]>=0.24.0
fastapi>=0.104.0
jinja2
python-multipart

# --- Banco de Dados ---
sqlmodel
aiosqlite
chromadb

# --- Web Crawler ---
requests
beautifulsoup4
lxml

# --- AI / LangChain ---
langchain
langchain-google-genai
langchain-core
langchain-text-splitters
ollama 

# --- Configuração ---
python-dotenv
pydantic

# --- Desenvolvimento & Testes ---
pytest>=8.4.0
pytest-mock
httpx